Split repository settings custom command views
